marinemap_requirements.txt
#############################################################################
# pip requirements file for marinemap/lingcod
#
# Automates the installation of most marinemap python dependencies
# To begin you need pip, virtualenv, hg, svn
#
# Usage:
# -----------------------------------------
# cd /usr/local
# sudo pip install -r marinemap_requirements.txt
# # OR if you want to upgrade existing installations
# sudo pip install --upgrade -r marinemap_requirements.txt
# # OR if you want to reference by URL
# sudo pip install --upgrade -r http://marinemap.googlecode.com/hg/marinemap_requirements.txt
# -----------------------------------------
# 
# You can also use this with virtualenv
# -----------------------------------------
# virtualenv MyMarineMap
# pip -E MyMarineMap install -r marinemap_requirements.txt
# cd MyMarineMap
# source bin/activate
# deactivate
# -----------------------------------------
#
# Note that there are several hairy requirements that aren't handled 
# by this file; these must be installed globally the hard way:
#
# GDAL >= 1.6 : just copy the gdal.pth file into your local site packages dir
# Mapnik == 0.7.1 : just copy the entire mapnik dir into your local site packages dir
# PIL : symlink the PIL directory to site-packages dir
# Postgis : set up according to the geodjango and marinemap docs
#############################################################################

-e svn+http://code.djangoproject.com/svn/django/trunk#egg=Django
BeautifulSoup>=3.1.0.1
Sphinx>=0.6.3
django-compress>=1.0.1
django-maintenancemode>=0.9.2
-e hg+http://bitbucket.org/ubernostrum/django-registration/#egg=django-registration
-e svn+http://feedvalidator.googlecode.com/svn/trunk/feedvalidator/src/#egg=feedvalidator
-e hg+https://bitbucket.org/springmeyer/djmapnik#egg=djmapnik
elementtree
xlwt>=0.7.2
networkx==1.1
boto
django-celery>=2.2.4
django-unittest-depth>=0.6
python-openid>=2.2.5
django-picklefield>=0.1.9
python-cjson>=1.0.5
epydoc>=3.0.1
docutils>=0.5
South>=0.7
psycopg2>=2.0.12
wsgiref>=0.1.2
numpy>=1.3.0
unittest-xml-reporting>=1.2
django-kombu>=0.9.2
